requirements.txt
blessings==1.5
Flask==0.10.1
monk==0.11.2
prettytable==0.7.2
python-dateutil==2.2
PyYAML==3.10
pyxdg==0.25

# for curses TUI
urwid==1.1.1

# for pim.web.providers.ttlbooks
rdflib==3.2.1
